+++ CHANGES.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+#### Version 2.0.2
+* Implement Social Network share buttons.
+* Bugfix: Titles not displayed for photos in masonry gallery.
+* Improve title display for the Dynamic Pages
+
 #### Version 2.0.1
 * Problem with .gitignore